# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POS / POS Referral Onboarding API",
    "description": "Backend for the digital onboarding flow (link generation, form, documents, approvals). Login/auth is provided by an existing system and is not part of this service.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://grid.leadsgenerations.in/recrument-analysis",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/api/v1/admin/onboarding": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Onboarding",
        "operationId": "list_onboarding_api_v1_admin_onboarding_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OnboardingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OnboardingListItem"
                  },
                  "title": "Response List Onboarding Api V1 Admin Onboarding Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/onboarding/{record_id}/review": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Review Onboarding",
        "operationId": "review_onboarding_api_v1_admin_onboarding__record_id__review_post",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnboardingListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          },
          "status": {
            "$ref": "#/components/schemas/OnboardingStatus"
          },
          "raised_by": {
            "type": "string",
            "title": "Raised By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "number",
          "user_type",
          "status",
          "raised_by",
          "created_at"
        ],
        "title": "OnboardingListItem"
      },
...
      "OnboardingStatus": {
        "type": "string",
        "enum": [
          "LINK_GENERATED",
          "FORM_IN_PROGRESS",
          "SUBMITTED",
          "UNDER_REVIEW",
          "SEND_BACK",
          "REJECTED",
          "APPROVED",
          "ACTIVE"
        ],
        "title": "OnboardingStatus"
      },
...
      "ReviewAction": {
        "type": "string",
        "enum": [
          "APPROVE",
          "REJECT",
          "SEND_BACK"
        ],
        "title": "ReviewAction"
      },
      "ReviewRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ReviewAction"
          },
          "reviewed_by": {
            "type": "string",
            "title": "Reviewed By"
          }
        },
        "type": "object",
        "required": [
          "action",
          "reviewed_by"
        ],
        "title": "ReviewRequest"
      },
      "ReviewResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OnboardingStatus"
          },
          "new_onboarding_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Onboarding Url"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ReviewResponse"
      },
...
      "UserType": {
        "type": "string",
        "enum": [
          "POS",
          "POS_REFERRAL"
        ],
        "title": "UserType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}